Move events and debug to own files
